Updated Sockets
Trying a new approach where I reopen the socket every time I write to the client, this is the only way I have found (thus far with my limited knowledge of sockets) to enable multiple lines to be written to the client from the server. Will be adding commands to drive the robot (Chandler) within the client application to test how fast the commands can be sent (if its fast enough this method should have no issues).